+++ src/main/java/codesqaud/app/config/ServletContextInitializer.java
@@ -1,6 +1,10 @@
 package codesqaud.app.config;
 
 import codesqaud.app.dao.*;
+import codesqaud.app.dao.article.ArticleDao;
+import codesqaud.app.dao.article.DbArticleDao;
+import codesqaud.app.dao.user.DbUserDao;
+import codesqaud.app.dao.user.UserDao;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletContextEvent;
 import jakarta.servlet.ServletContextListener;
