+++ src/main/java/codesqaud/app/dao/article/ArticleDao.java
@@ -1,5 +1,6 @@
-package codesqaud.app.dao;
+package codesqaud.app.dao.article;
 
+import codesqaud.app.dao.CommonDao;
 import codesqaud.app.model.Article;
 
 public interface ArticleDao extends CommonDao<Article, Long> {
