+++ src/test/java/codesqaud/app/dao/InMemoryUserDaoTest.java
@@ -1,9 +1,9 @@
 package codesqaud.app.dao;
 
+import codesqaud.app.dao.user.InMemoryUserDao;
 import codesqaud.app.exception.HttpException;
 import codesqaud.app.model.User;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
